% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnca.R
\name{ggnca}
\alias{ggnca}
\title{ggplot for pharmacokinetic concentration-time curve}
\usage{
ggnca(concData, colSubj = "Subject", colTime = "Time", colConc = "conc")
}
\arguments{
\item{concData}{concentration data table}

\item{colSubj}{column name for subject ID}

\item{colTime}{column name for time}

\item{colConc}{column name for concentration}
}
\value{
ggplot2 figures
}
\description{
This draws pharmacokinetic concentration-time curve with ggplot2 packages.
}
\examples{
ggnca(concData = Theoph, colSubj = "Subject", colTime = "Time", colConc = "conc")

}
