% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnca.R
\name{plot_conc}
\alias{plot_conc}
\title{concentration-time curves}
\usage{
plot_conc(df)
}
\arguments{
\item{df}{concentration data table}

\item{colSubj}{column name for subject ID}

\item{colTime}{column name for time}

\item{colConc}{column name for concentration}
}
\description{
This draws pharmacokinetic concentration-time curve with ggplot2 packages.
}
